pyproject.toml
[project]
name = "fckunigotomkt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
]
